Move Status to game module
