+++ src/glicko2/game.rs
@@ -1,8 +1,7 @@
-/*! 
+/*!
 Provides functions to handle a single one on one game and update ratings accordingly
 */
-use crate::glicko2::algorithm;
-use crate::glicko2::rating::{Rating, Status};
+use crate::glicko2::{algorithm, constants, rating::Rating};
 
 /// Updates ratings for two teams
 /// If the game was a draw, pass `drawn` as `true`.
@@ -10,9 +9,12 @@ use crate::glicko2::rating::{Rating, Status};
 /// # Example
 ///
 /// ```
-/// let mut rating_1 = glicko2::rating::Rating::new();
-/// let mut rating_2 = glicko2::rating::Rating::new();
-/// let (rating_1, rating_2) = glicko2::game::compete(rating_1, rating_2, false);
+/// use glicko2::{rating::Rating, game};
+///
+/// let mut rating_1 = Rating::new();
+/// let mut rating_2 = Rating::new();
+///
+/// let (rating_1, rating_2) = game::compete(rating_1, rating_2, false);
 /// ```
 pub fn compete(mut winner: Rating, mut loser: Rating, drawn: bool) -> (Rating, Rating) {
     // drawn is false if Team 1 beat Team 2
@@ -31,9 +33,12 @@ pub fn compete(mut winner: Rating, mut loser: Rating, drawn: bool) -> (Rating, R
 /// # Example
 ///
 /// ```
-/// let mut rating_1 = glicko2::rating::Rating::new();
-/// let mut rating_2 = glicko2::rating::Rating::new();
-/// let odds = glicko2::game::odds(rating_1, rating_2);
+/// use glicko2::{rating::Rating, game};
+///
+/// let mut rating_1 = Rating::new();
+/// let mut rating_2 = Rating::new();
+///
+/// let odds = game::odds(rating_1, rating_2);
 /// ```
 pub fn odds(mut rating1: Rating, mut rating2: Rating) -> f64 {
     rating1.scale_down();
@@ -53,9 +58,12 @@ pub fn odds(mut rating1: Rating, mut rating2: Rating) -> f64 {
 /// # Example
 ///
 /// ```
-/// let mut rating_1 = glicko2::rating::Rating::new();
-/// let mut rating_2 = glicko2::rating::Rating::new();
-/// let quality = glicko2::game::quality(rating_1, rating_2);
+/// use glicko2::{rating::Rating, game};
+///
+/// let mut rating_1 = Rating::new();
+/// let mut rating_2 = Rating::new();
+///
+/// let quality = game::quality(rating_1, rating_2);
 /// ```
 pub fn quality(rating1: Rating, rating2: Rating) -> f64 {
     // 1.0 if perfect match
@@ -63,4 +71,32 @@ pub fn quality(rating1: Rating, rating2: Rating) -> f64 {
     let expected_score_2 = odds(rating2, rating1);
     let advantage = expected_score_1 - expected_score_2; // Advantage team 1 has over team 2
     1.0 - advantage.abs()
+}
+
+/// Enum representing the Glicko2 values for match outcomes
+#[derive(Debug)]
+pub enum Status {
+    Win,
+    Draw,
+    Loss,
+}
+
+impl Status {
+    /// Gets the constant float value associated with each outcome
+    ///
+    /// # Example
+    ///
+    /// ```
+    /// use glicko2::game::Status;
+    ///
+    /// let loss = Status::Loss;
+    /// let loss_val = loss.val();
+    /// ```
+    pub fn val(&self) -> f64 {
+        match self {
+            Status::Win => constants::WIN,
+            Status::Draw => constants::DRAW,
+            Status::Loss => constants::LOSS,
+        }
+    }
 }
